Front Vue: Player: Add composable for browser metadata
